refactor: add auto-refresh feature and update SetupScopeGrid component in setup pages
