+++ otterscale-web/src/lib/components/setup/index.ts
@@ -1,1 +1,2 @@
 export { default as SetupScopeGrid } from "./setup-scope-grid.svelte";
+export { default as SetupScope } from "./setup-scope.svelte";
